Move to new architecture in which the hypernetwork simply wraps the ordinary net.
The .forward method of the hypernetwork will overwrite the parameters
of the ordinary network. And the .backward method will carry the
gradients from the net back into the hypernetwork.

Implemented a Dummy hypernetwork that merely mirrors the parameters of
the network. This is useful for testing, as it should yield the same
behavior is if training the underlying network.